[refactor] change code style in commandservice
